Move plugin configuration load to AgentConfiguration
